Moved code upwards
