+++ layers/Engine/packages/component-model/src/GraphQLParser/ExtendedSpec/Parser/Parser.php
@@ -5,20 +5,14 @@ declare(strict_types=1);
 namespace PoP\ComponentModel\GraphQLParser\ExtendedSpec\Parser;
 
 use PoP\ComponentModel\DirectiveResolvers\MetaDirectiveResolverInterface;
-use PoP\ComponentModel\HelperServices\QueryHelperServiceInterface;
 use PoP\ComponentModel\Registries\MetaDirectiveRegistryInterface;
 use PoP\GraphQLParser\ExtendedSpec\Parser\AbstractParser;
-use PoP\GraphQLParser\ExtendedSpec\Parser\Ast\ArgumentValue\DynamicVariableReference;
 use PoP\GraphQLParser\Spec\Parser\Ast\Argument;
-use PoP\GraphQLParser\Spec\Parser\Ast\ArgumentValue\Variable;
-use PoP\GraphQLParser\Spec\Parser\Ast\ArgumentValue\VariableReference;
 use PoP\GraphQLParser\Spec\Parser\Ast\Directive;
-use PoP\GraphQLParser\Spec\Parser\Location;
 
 class Parser extends AbstractParser
 {
     private ?MetaDirectiveRegistryInterface $metaDirectiveRegistry = null;
-    private ?QueryHelperServiceInterface $queryHelperService = null;
 
     final public function setMetaDirectiveRegistry(MetaDirectiveRegistryInterface $metaDirectiveRegistry): void
     {
@@ -27,14 +21,6 @@ class Parser extends AbstractParser
     final protected function getMetaDirectiveRegistry(): MetaDirectiveRegistryInterface
     {
         return $this->metaDirectiveRegistry ??= $this->instanceManager->getInstance(MetaDirectiveRegistryInterface::class);
-    }
-    final public function setQueryHelperService(QueryHelperServiceInterface $queryHelperService): void
-    {
-        $this->queryHelperService = $queryHelperService;
-    }
-    final protected function getQueryHelperService(): QueryHelperServiceInterface
-    {
-        return $this->queryHelperService ??= $this->instanceManager->getInstance(QueryHelperServiceInterface::class);
     }
 
     protected function isMetaDirective(string $directiveName): bool
@@ -69,21 +55,5 @@ class Parser extends AbstractParser
         /** @var MetaDirectiveResolverInterface */
         $metaDirectiveResolver = $this->getMetaDirectiveResolver($directive->getName());
         return $metaDirectiveResolver->getAffectDirectivesUnderPosArgumentDefaultValue();
-    }
-
-    protected function createVariableReference(
-        string $name,
-        ?Variable $variable,
-        Location $location,
-    ): VariableReference {
-        if ($this->getQueryHelperService()->isDynamicVariableReference($name, $variable)) {
-            return new DynamicVariableReference($name, $variable, $location);
-        }
-
-        return parent::createVariableReference(
-            $name,
-            $variable,
-            $location,
-        );
     }
 }
