+++ layers/Engine/packages/graphql-parser/src/ExtendedSpec/Parser/AbstractParser.php
@@ -5,12 +5,16 @@ declare(strict_types=1);
 namespace PoP\GraphQLParser\ExtendedSpec\Parser;
 
 use PoP\ComponentModel\Feedback\FeedbackItemResolution;
+use PoP\ComponentModel\HelperServices\QueryHelperServiceInterface;
 use PoP\GraphQLParser\Component;
 use PoP\GraphQLParser\ComponentConfiguration;
 use PoP\GraphQLParser\Exception\Parser\InvalidRequestException;
+use PoP\GraphQLParser\ExtendedSpec\Parser\Ast\ArgumentValue\DynamicVariableReference;
 use PoP\GraphQLParser\ExtendedSpec\Parser\Ast\MetaDirective;
 use PoP\GraphQLParser\FeedbackItemProviders\GraphQLExtendedSpecErrorFeedbackItemProvider;
 use PoP\GraphQLParser\Spec\Parser\Ast\Argument;
+use PoP\GraphQLParser\Spec\Parser\Ast\ArgumentValue\Variable;
+use PoP\GraphQLParser\Spec\Parser\Ast\ArgumentValue\VariableReference;
 use PoP\GraphQLParser\Spec\Parser\Ast\Directive;
 use PoP\GraphQLParser\Spec\Parser\Location;
 use PoP\GraphQLParser\Spec\Parser\Parser as UpstreamParser;
@@ -18,6 +22,17 @@ use PoP\Root\App;
 
 abstract class AbstractParser extends UpstreamParser implements ParserInterface
 {
+    private ?QueryHelperServiceInterface $queryHelperService = null;
+    
+    final public function setQueryHelperService(QueryHelperServiceInterface $queryHelperService): void
+    {
+        $this->queryHelperService = $queryHelperService;
+    }
+    final protected function getQueryHelperService(): QueryHelperServiceInterface
+    {
+        return $this->queryHelperService ??= $this->instanceManager->getInstance(QueryHelperServiceInterface::class);
+    }
+
     /**
      * Replace `Directive` with `MetaDirective`, and nest the affected
      * directives inside.
@@ -234,5 +249,21 @@ abstract class AbstractParser extends UpstreamParser implements ParserInterface
         Location $location,
     ): MetaDirective {
         return new MetaDirective($name, $arguments, $nestedDirectives, $location);
+    }
+
+    protected function createVariableReference(
+        string $name,
+        ?Variable $variable,
+        Location $location,
+    ): VariableReference {
+        if ($this->getQueryHelperService()->isDynamicVariableReference($name, $variable)) {
+            return new DynamicVariableReference($name, $variable, $location);
+        }
+
+        return parent::createVariableReference(
+            $name,
+            $variable,
+            $location,
+        );
     }
 }
